+++ composer.json
@@ -22,8 +22,7 @@
         "doctrine/inflector": "^1.1",
         "dragonmantank/cron-expression": "^2.0",
         "egulias/email-validator": "^2.1.10",
-        "league/commonmark": "^1.1",
-        "league/commonmark-ext-table": "^2.1",
+        "league/commonmark": "^1.3",
         "league/flysystem": "^1.0.8",
         "monolog/monolog": "^1.12|^2.0",
         "nesbot/carbon": "^2.0",
