+++ src/Illuminate/Foundation/Testing/Concerns/MakesHttpRequests.php
@@ -26,6 +26,13 @@ trait MakesHttpRequests
     protected $defaultCookies = [];
 
     /**
+     * Additional cookies will not be encrypted for the request.
+     *
+     * @var array
+     */
+    protected $unencryptedCookies = [];
+
+    /**
      * Additional server variables for the request.
      *
      * @var array
@@ -168,6 +175,33 @@ trait MakesHttpRequests
     public function withCookie(string $name, string $value)
     {
         $this->defaultCookies[$name] = $value;
+
+        return $this;
+    }
+
+    /**
+     * Define additional cookies will not be encrypted before sending with the request.
+     *
+     * @param  array  $cookies
+     * @return $this
+     */
+    public function withUnencryptedCookies(array $cookies)
+    {
+        $this->unencryptedCookies = array_merge($this->unencryptedCookies, $cookies);
+
+        return $this;
+    }
+
+    /**
+     * Add a cookie will not be encrypted before sending with the request.
+     *
+     * @param  string  $name
+     * @param  string  $value
+     * @return $this
+     */
+    public function withUnencryptedCookie(string $name, string $value)
+    {
+        $this->unencryptedCookies[$name] = $value;
 
         return $this;
     }
@@ -527,12 +561,12 @@ trait MakesHttpRequests
     protected function prepareCookiesForRequest()
     {
         if (! $this->encryptCookies) {
-            return $this->defaultCookies;
+            return array_merge($this->defaultCookies, $this->unencryptedCookies);
         }
 
         return collect($this->defaultCookies)->map(function ($value) {
             return encrypt($value, false);
-        })->all();
+        })->merge($this->unencryptedCookies)->all();
     }
 
     /**
