+++ src/Illuminate/Mail/composer.json
@@ -19,8 +19,7 @@
         "illuminate/container": "^6.0",
         "illuminate/contracts": "^6.0",
         "illuminate/support": "^6.0",
-        "league/commonmark": "^1.1",
-        "league/commonmark-ext-table": "^2.1",
+        "league/commonmark": "^1.3",
         "psr/log": "^1.0",
         "swiftmailer/swiftmailer": "^6.0",
         "tijsverkoyen/css-to-inline-styles": "^2.2.1"
