+++ src/Illuminate/Redis/Connectors/PhpRedisConnector.php
@@ -98,10 +98,6 @@ class PhpRedisConnector implements Connector
                 $client->setOption(Redis::OPT_READ_TIMEOUT, $config['read_timeout']);
             }
 
-            if (! empty($options['serializer'])) {
-                $client->setOption(Redis::OPT_SERIALIZER, $options['serializer']);
-            }
-
             if (! empty($config['scan'])) {
                 $client->setOption(Redis::OPT_SCAN, $config['scan']);
             }
@@ -160,16 +156,12 @@ class PhpRedisConnector implements Connector
                 $client->setOption(RedisCluster::OPT_PREFIX, $options['prefix']);
             }
 
-            if (! empty($options['serializer'])) {
-                $client->setOption(RedisCluster::OPT_SERIALIZER, $options['serializer']);
+            if (! empty($options['scan'])) {
+                $client->setOption(RedisCluster::OPT_SCAN, $options['scan']);
             }
 
-            if (! empty($config['scan'])) {
-                $client->setOption(RedisCluster::OPT_SCAN, $config['scan']);
-            }
-
-            if (! empty($config['failover'])) {
-                $client->setOption(RedisCluster::OPT_SLAVE_FAILOVER, $config['failover']);
+            if (! empty($options['failover'])) {
+                $client->setOption(RedisCluster::OPT_SLAVE_FAILOVER, $options['failover']);
             }
         });
     }
